setting: 기존 파일 제거 및 타입 이동
