--- src/apis/store/deleteStoreApi.ts
@@ -1,13 +1,0 @@
-import apiClient from "../apiClient.ts";
-
-/**
- * 매장 삭제
- */
-const deleteStore = async <T = { msg: string }>(): Promise<T> => {
-  const res = await apiClient.delete<T>("/store", {
-    withCredentials: true,
-  });
-  return res.data;
-};
-
-export { deleteStore };
