--- src/apis/store/getMyStoreApi.ts
@@ -1,21 +1,0 @@
-import apiClient from "../apiClient.ts";
-import type { IStore } from "../../types/store.d.ts";
-
-interface MyStoreApiData {
-  store: IStore;
-  comments: object[]; // CommentType 받은 이후 수정 필요
-}
-
-/**
- * 등록한 매장(내 가게) 정보
- */
-const getMyStore = async <
-  T = { msg: string } & MyStoreApiData,
->(): Promise<T> => {
-  const res = await apiClient.get<T>(`/store`, {
-    withCredentials: true,
-  });
-  return res.data;
-};
-
-export { getMyStore };
